# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Human Link",
    "version": "3.0.0",
    "description": "Généré depuis les routeurs Express et les schémas zod partagés (shared/src/index.ts). Authentification par cookie httpOnly (`hl_at`). Les écritures (POST, PUT, PATCH, DELETE) exigent aussi l’en-tête `X-CSRF-Token` recopiant le cookie `hl_csrf` (obtenu par GET /api/auth/csrf). Les dates sont au format AAAA-MM-JJ."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/contact": {
      "get": {
        "tags": [
          "contact"
        ],
        "operationId": "get_contact",
        "summary": "Messages reçus",
        "description": "Rôles autorisés : admin. Authentification requise (cookie de session).",
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "userId": {
                            "type": "string"
                          },
                          "nom": {
                            "type": "string"
                          },
                          "prenom": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "telephone": {
                            "type": "string"
                          },
                          "sujet": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "email",
                          "sujet",
                          "message",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "403": {
            "description": "Rôle insuffisant"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true,
        "x-roles": [
          "admin"
        ]
      },
      "post": {
        "tags": [
          "contact"
        ],
        "operationId": "post_contact",
        "summary": "Envoyer un message à l’équipe RH",
        "description": "Ouvert aux visiteurs ; limité en débit ; champ piège `website` contre les robots. Public.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nom": {
                    "type": "string",
                    "maxLength": 80
                  },
                  "prenom": {
                    "type": "string",
                    "maxLength": 80
                  },
                  "email": {
                    "type": "string",
                    "maxLength": 254,
                    "format": "email",
                    "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "telephone": {
                    "type": "string",
                    "maxLength": 30
                  },
                  "sujet": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 160
                  },
                  "message": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 4000
                  },
                  "website": {
                    "type": "string",
                    "maxLength": 200
                  }
                },
                "required": [
                  "sujet",
                  "message"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-documented": true
      }
    },
    "/api/contact/{id}": {
      "delete": {
        "tags": [
          "contact"
        ],
        "operationId": "delete_contact_id",
        "summary": "Supprimer un message",
        "description": "Rôles autorisés : admin. Authentification requise (cookie de session).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "403": {
            "description": "Rôle insuffisant"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true,
        "x-roles": [
          "admin"
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hl_at"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}